test user implemented

+++ backend/controllers/recipeControllers.js
@@ -13,16 +13,21 @@ export const createRecipe = async (req, res) => {
     throw new ExpressError("No data received", 404);
   }
   //cloudinary upload
-  if (req.file) {
-    console.log(req.file);
-    const response = await cloudinary.v2.uploader.upload(req.file.path);
-    await fs.unlink(req.file.path);
+  if (req.body.avatarUrl) {
+    //check avatarUrl in the req.body from the forms(we used state to submit)
+    const response = await cloudinary.v2.uploader.upload(
+      req.body.avatarUrl,
+      req.body.avatarPublicId
+    ); //path of the image file in the public storage
     req.body.avatarUrl = response.secure_url;
     req.body.avatarPublicId = response.public_id;
   }
   //UPDATE: adding a created by property
+  //Update created a new property in the RecipeSchema that will hold the value of the owner of the recipe (coming from the req.user.userName upon logging in).
   req.body.createdBy = req.user.userId;
+  req.body.author = req.user.userName;
   const newRecipe = await RecipeModel.create(req.body);
+  console.log(newRecipe);
   if (!newRecipe) {
     throw ExpressError("No recipe created", 400);
   }
@@ -30,9 +35,10 @@ export const createRecipe = async (req, res) => {
 };
 
 //all recipes
+//Update: changed finding all recipe instead of the ones created by the logged user.
 export const getAllRecipes = async (req, res) => {
-  const allRecipes = await RecipeModel.find({ createdBy: req.user.userId }); //find all recipe of logged user, returns an array
-  console.log(allRecipes);
+  const allRecipes = await RecipeModel.find({}); //find all recipe, returns an array
+  // console.log(allRecipes);
   if (allRecipes.length === 0) {
     throw new ExpressError("No recipes found", 404);
   }
@@ -48,9 +54,10 @@ export const getSingleRecipe = async (req, res) => {
     throw new ExpressError("Recipe not found", 404);
   }
   //limiting access of a single recipe depending if user is admin or the author.
-  if (req.user.role !== "admin" && singleRecipe.createdBy !== req.user.userId) {
-    throw new ExpressError("user not authorized");
-  }
+  // if (req.user.role !== "admin" && singleRecipe.createdBy !== req.user.userId) {
+  //   throw new ExpressError("user not authorized");
+  // }
+
   res.status(200).json({ message: "Recipe found", singleRecipe });
 };
 
@@ -59,7 +66,24 @@ export const editRecipe = async (req, res) => {
   if (!req.body) {
     throw new ExpressError("No data found", 404);
   }
+
   const { id } = req.params;
+  if (req.body.avatarUrl) {
+    const response = await cloudinary.v2.uploader.upload(req.body.avatarUrl); //cloudinary upload API
+    req.body.avatarUrl = response.secure_url;
+    req.body.avatarPublicId = response.public_id;
+  }
+
+  //searching the specific recipe using params to check whether it contains a avatarPublicId property, which we will then destroy using cloudinary destroy API to avoid uploading a new image to cloudinary everytime
+  const oldRecipe = await RecipeModel.findById(id);
+  if (oldRecipe.avatarPublicId) {
+    await cloudinary.v2.uploader.destroy(oldRecipe.avatarPublicId);
+  }
+
+  //limiting the access to editing recipe
+  if (req.user.role !== "admin" && oldRecipe.createdBy !== req.user.userId) {
+    throw new ExpressError("User not authorized", 400);
+  }
   const editedRecipe = await RecipeModel.findByIdAndUpdate(id, req.body, {
     new: true,
   });
@@ -72,6 +96,11 @@ export const editRecipe = async (req, res) => {
 //delete recipe
 export const deleteRecipe = async (req, res) => {
   const { id } = req.params;
+  //limit deleting of recipe to authors only
+  const foundRecipe = await RecipeModel.findById(id);
+  if (req.user.role !== "admin" && foundRecipe.createdBy !== req.user.userId) {
+    throw new ExpressError("User not authorized", 400);
+  }
   const deletedRecipe = await RecipeModel.findByIdAndDelete(id);
   if (!deletedRecipe) {
     throw new ExpressError("cannot delete recipe", 404);
